app/backend/pyproject.toml
[project]
name = "marketpl-backend"
version = "0.1.0"
description = "MarketPl.ai backend service"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "pymongo==4.13.2",
  "python-dotenv==1.1.1",
]
